tdf#150806 ODT export: fix fallback svg:width/height for zero layout size
The reported problem was that once tracked changes is enabled and an
image is deleted, its aspect ratio changes on save.

What happens here is that the image's frame is marked as a redline
deletion, so when the ODT export writes the file, it'll have a zero
size. But this zero size is nothing we should take into account on save,
XMLTextParagraphExport::addTextFrameAttributes() expected to always have
a valid layout size if the LayoutSize property is available.

Fix the problem by adding one more condition where we ignore the layout
size, which was already there since commit
51487f3687405c918c5dadf6b946d02824488450 (tdf#150990 ODT export: fix
zero layout size of scale/scale images, 2022-09-20).

Note that this way the written side will be the original one, so only
the aspect ratio will be correct, but that was accepted as a compromise
earlier.

Change-Id: I108dc2428ce0cb07a044bfef216575fdba19267b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/146043
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>