sw: add document model for multi-page fly frames
Currently we can have multi-page tables and fly frames (where the
content wraps around the fly frame), but not both at the same time.

Allowing multi-page fly frames in general would probably go too far, but
handling this in case the fly frame only contains a table would make
sense to support Word's multi-page floating tables.

Fix this by adding a new SwFormatFlySplit property on fly frames: this
is meant to be enabled when the fly frame is part of a multi-page
floating table.

Unlike table split, this is off by default. Some logic will be needed
later to enable this in case the fly frame originates from a Word
floating table. (And obvously this is just the doc model, everything
else still needs doing.)

Change-Id: Id92130c684016cd7e7153e15b0ec6821e4471879
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/146095
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins