android: Use actual Android API level for minSdkVersion, nss
When an Android API level is explicitly set with the
`--with-android-api-level` switch introduced in

    commit 09f7c0f7b0e5db7c42e439d1c8448a77a5b8d40d
    Date:   Thu Jul 4 09:06:49 2019 +0200

        android: Allow specification of the API level.

, use that for the minSdkVersion for the Android Viewer app
and the API level for the NSS build, rather than leaving
that hard-coded at API level 16.
Building with a newer API level means that the
app won't run on devices with older API levels.
With this in place, this will be recognized at
install time (installation will fail: INSTALL_FAILED_OLDER_SDK)
rather than crashing at run time.

Change-Id: Id6047b768d265b965696f3a3161d7828e5f3696e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/146127
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>