android: Support NDK 24.x and 25.x, use NDK 25 for Jenkins
NDK 24 dropped support for API versions < 19,
quoting [1]:

> * Jelly Bean (APIs 16, 17, and 18) is no longer supported. The minimum OS
>   supported by the NDK is KitKat (API level 19).

Therefore, use API level 19 for these (and unknown) NDKs unless
anything greater has explicitly been specified using the
`--with-android-api-level=<VERSION>` autogen switch
(or is default for the architecture, s. above).

Update the Jenkins config to use NDK 25 for the Android
builds as discussed in the ESC meeting on 2023-01-26 [2].

[1] https://developer.android.com/ndk/downloads/revision_history
[2] https://lists.freedesktop.org/archives/libreoffice/2023-January/089878.html

Change-Id: Ib8e65f433ee89ff1bc12432722570bf8f9f7ed85
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/146135
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>