sw: inherit SwFlyAtContentFrame from SwFlowFrame
Checking how e.g. section frames are split in SwFrame::GetNextLeaf(), we
call into SwFrame::GetNextSctLeaf() which then calls a dedicated
SwSectionFrame ctor that connects the master with its follow.

Add infrastructure to be able to do the same for fly at-content frames.
(At-fly anchored and as-char anchored frames are not relevant for the
broader multi-page floating table picture. Also, SwFrame::GetLeaf()
wants to work with a non-abstract type, so allocating SwFlyFrame there
would make no sense, so inherit SwFlyAtContentFrame and not SwFlyFrame
from SwFlowFrame.)

This is meant to be a minimal change, that can be unconditional, seeing
that the only dynamic_cast to SwFlowFrame in RestoreContent() only does
something if the frame has a follow, which is never the case at the
moment.

Towards an initial layout for multi-page fly frames.

Change-Id: I8ea930bed4bbddb96b181796ab893e810ba9dcd4
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/146160
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins