use ThemeColorChanger also in svx, when changing theme for SdrPage
This refactors the code in the Theme class to remove code that is
responsible for changing the color for the SdrPage and adds an
implementation of IThemeColorChanger - svx::ThemeColorChanger,
that is performing the same task from now on.

The svx::ThemeColorChanger is also partially used in Writer by
sw::ThemeColorChanger, when it needs to change the colors in the
common (Sdr) objects.

Change-Id: I39cfa278d31bdd802a5bf36eeaf6d472c8013dba
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/146173
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>