sw: add an SwFlyAtContentFrame ctor that creates a follow fly frame
This is meant to be a called by a future SwFrame::GetNextFlyLeaf()
(similar to the existing SwFrame::GetNextSctLeaf()), which still has to
be added.

What an anchor should be for a follow fly frame is a bit unclear to me,
let's go with an explicit nullptr for now and we can clarify that later.

Towards an initial layout for multi-page fly frames.

Change-Id: I2429f47aef90dfe61340bfe244353dcefeb4ef1f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/146244
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins