tdf#153296 improved markup for VML shapetype of Fontwork
In the current version of 'vml-shape-types' file the markup for the
Fontwork shapes is missing or is so, that Word does not accept it as
WordArt markup. So Word treats it as custom geometry and saves it that
way. But LO cannot import custom geometry for WordArt.

The patch changes it so, that for Fontwork shapes a dedicated markup
map is used with markup, that Word accepts as WordArt. Non-Fontwork
shapes still use the 'vml-shape-types' file for VML export.

I have added two unit tests. But a true test is not possible, because
for that you would need to open and save the files with MS Word.

This patch solves at the same time some of the handle problems for the
special case of Fontwork shapes, see tdf#152913.

Change-Id: Icc5afb88c58b644d25b3cf8d415e5aa2d6a7b19e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/146311
Tested-by: Jenkins
Reviewed-by: Regina Henschel <rb.henschel@t-online.de>