tdf#153083 writerfilter: import locale-dependent TOC \t style names, 2
There was a problem with commit aa9d326d76252a3fe1e3f999ba3b32a7f65c3fec
while it now worked in Writer, the roundtripped DOCX with the en-US
built-in style name in the TOC field no longer worked in Word in the
original locale (but started to work in en-US locale).

Let's try a slightly different approach: use the same style name as-is
from the TOC field, and if it doesn't exist as a style, then clone the
en-US built-in style with that name, and at the end of the import,
iterate all paragraphs in the document and replace any application of the
en-US style with the localised style.

So both the en-US style and the localised style exist, and for
references that aren't fields (e.g. basedOn in other styles) it should
hopefully not matter which one is referenced.

The DOCX exported from Writer now has a TOC field that works in Word in
the original locale.

Change-Id: Ibcc3c5899e31295b5704ebefb548f40b67eda9bf
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/146414
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>