sw: avoid unwanted initial content in split/follow fly frames
If the fly on the first gets split, SwFrame::GetNextFlyLeaf() will
create a follow fly frame. Don't fill this one with content, we'll
instead want to move some of the content of the master here. (Which is
not yet working.)

InsertCnt() is called *really* early, so we can't check if this frame
has a master -- just pass a bool around instead.

This fixes the unwanted first paragraph in the follow fly frame at
least. (The one that fits the master already.)

Towards an initial layout for multi-page fly frames.

Change-Id: If364fdaf7595fbc5addbc6b3b3b0092e958c5380
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/146539
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins