sw: implement SwFrame::GetPrevFlyLeaf()
This is much easier than the "next" case, since here we can nullptr when
there is no previous leaf instead of creating one.

With this, if the MoveBwd() call in SwContentFrame::MakeAll() is
disabled, some simple fly frame can be split across 2 pages if it's
positioned in a way that the rectangle of the fly would not fit into the
body frame without changing the position, in SW_FORCE_FLY_SPLIT=1 mode.

Towards an initial layout for multi-page fly frames.

Change-Id: I3b6ba85ad66117b5595a113d688e2fcb97bcf745
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/146571
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins