Fix thinko when to sleep
<https://ci.libreoffice.org/job/lo_ubsan/2672/> still failed with the same
symptoms that 0109e369543377e4e4f93e2451b6086b768686f2 "A better fix for the
stale TreeListEntryUIObject issue" had been supposed to fix.  And of course, it
is the

>                 xInsert.executeAction("CLICK", tuple())

calls that actually trigger m_aUpdateTimer.Start() in
SwChildWinWrapper::ReInitDlg, while the upfront m_aUpdateTimer.SetTimeout(200)
in the SwChildWinWrapper ctor by itself is not relevant for the point in time
where the timer actually starts (and where the Python code thus has to start
waiting, too).

Change-Id: I2b4fe7dadd470c62e7d63594a76d258faeac5906
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/146588
Tested-by: Jenkins
Reviewed-by: Stephan Bergmann <sbergman@redhat.com>