tdf#153083 writerfilter,sw: DOCX locale-dependent TOC \t style names, 3
The import crashed on fdo85740-1.docx with an unhandled exception.

This is because it mapped the TOC style "Table of Figures" from "table
of figures", but the ApplyStyleSheetsImpl() actually uses the converted
name "Drawing" instead, so "Table of Figures" didn't exist.

(regression from commit a5835bdd466f75678601428fcf0147d665c4f8e2)

It turns out that once that is fixed, the DOCX export needs to convert
"Drawing" back to "Table of Figures" in the TOC field.

Change-Id: I4858c79dd74154b229b7568610c0b8ba7b3e2b6a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/146610
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>