tdf#153476 Disable MS telemetry in VsDevCmd.bat
MS telemetry inside VsDevCmd.bat causes problems for the autogen.sh
script which uses it to find the Visual Studio C/C++ standard library
via the environment variables that show the UCRT path and version.

VsDevCmd.bat is the "Developer Command Prompt for Visual Studio".

The telemtry is disabled by:

  set VSCMD_SKIP_SENDTELEMETRY=1

This change resolves the problems described in tdf#153476.

Change-Id: I8b77a2ca0c3577b3ba490723ab7901f463562578
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/146677
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>