tdf#153352 Use Promote/Demote in Tabbed UI instead of Increase/Decrease indent
As discussed in the ticket, the Tabbed UI of Impress/Draw uses the "Increase/Decrease indent" commands (.uno:DecrementIndent and .uno:IncrementIndent) which is confusing for the user.

When the user clicks this command, they want actually to Promote/Demote the list level, so instead we should use the commands .uno:OutlineLeft and .uno:OutlineRight here instead. This patch makes the Tabbed UI more coherent with the sidebar, which uses the correct commands.

Change-Id: I070b52e6fb31d348b1dd97c0ba28911f6e73590a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/146705
Tested-by: Jenkins
Tested-by: Heiko Tietze <heiko.tietze@documentfoundation.org>
Reviewed-by: Heiko Tietze <heiko.tietze@documentfoundation.org>
Reviewed-by: Rafael Lima <rafael.palma.lima@gmail.com>