editeng: split SvxLRSpaceItem into 1 class per attribute
The tricky part is that there are actually 2 different kinds of left
margins:

Those on entities that aren't paragraphs are simple, so class
SvxLeftMarginItem is simple.

Those on paragraphs are entangled with the first-line-offset,
because there is a left margin value (same as for everything else)
and a text-left-margin value, which includes a negative
first-line-offset; the text-left-margin value is exported as API
property so stored as-is in the item, and the left-margin is computed on
demand (based on corresponding first-line-offset) in
SvxTextLeftMarginItem::GetLeft().

Separate classes for these is a better separation than what commit
9150d607e2c9a22973246edff5559493de9a5e23 did.

(some functions aren't implemented because the SvxLRSpaceItem did
nothing for these cases; not sure if that's intended.)

Change-Id: Iae5cd4bf8167754eec65d2585430f5fe3d00839f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/147023
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>