tdf#153572 Implement .uno:ZoomPlus and .uno:ZoomMinus in the Basic IDE
This patch implements the commands .uno:ZoomPlus and .uno:ZoomMinus in the Basic IDE to allow the user to zoom in/out of the code editor. With these commands it is possible to create a keyboard shortcut for zooming in/out.

Notice that this patch only implements the commands, but it does not create a default shortcut, since defining which shortcut should be used needs to be discussed first.

Change-Id: I02866b9e731bf23181d0ecfc32f978ec110e33f9
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/146841
Tested-by: Jenkins
Tested-by: Heiko Tietze <heiko.tietze@documentfoundation.org>
Reviewed-by: Heiko Tietze <heiko.tietze@documentfoundation.org>